pyproject.toml
[tool.poetry]
name = "attendance-management"
version = "0.1.0"
description = "A secure attendance management system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.23"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"
cryptography = "^41.0.5"
aiosqlite = "^0.19.0"
alembic = "^1.12.1"
email-validator = "^2.1.0.post1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"